# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/shops": {
      "post": {
        "operationId": "ShopsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Taller creado (my_role = owner).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea un taller; quien lo crea queda como owner.",
        "tags": [
          "shops"
        ]
      }
    },
    "/v1/shops/mine": {
      "get": {
        "operationId": "ShopsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de talleres con mi rol en cada uno.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShopView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Talleres donde la cuenta actual es miembro activo (reclama antes sus invitaciones).",
        "tags": [
          "shops"
        ]
      }
    },
    "/v1/shops/{shopId}": {
      "get": {
        "operationId": "ShopsController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El taller con mi rol.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopView"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no soy miembro (no se distingue)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalle de un taller (solo miembros).",
        "tags": [
          "shops"
        ]
      }
    },
    "/v1/shops/{shopId}/members": {
      "get": {
        "operationId": "ShopMembersController_list",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de miembros.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShopMemberView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Miembros e invitaciones pendientes del taller (cualquier miembro).",
        "tags": [
          "shops"
        ]
      },
      "post": {
        "operationId": "ShopMembersController_invite",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Miembro activo o invitación pendiente (idempotente).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopMemberView"
                }
              }
            }
          },
          "403": {
            "description": "Solo el owner puede invitar."
          },
          "409": {
            "description": "Ya es miembro activo del taller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invita por email (solo owner). Con cuenta verificada entra al instante; si no, queda invitado y se le avisa por email.",
        "tags": [
          "shops"
        ]
      }
    },
    "/v1/shops/{shopId}/members/{memberId}": {
      "delete": {
        "operationId": "ShopMembersController_remove",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Quita un miembro o cancela una invitación (solo owner; al owner no).",
        "tags": [
          "shops"
        ]
      }
    },
...
    "/v1/shops/{shopId}/work-orders/{workOrderId}/invoice": {
      "post": {
        "operationId": "InvoicesController_issue",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Factura FAC-XXXX emitida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            }
          },
          "409": {
            "description": "OT no completada, ya facturada, o sin líneas aprobadas."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Factura una OT completada (snapshot inmutable de las líneas aprobadas). OT → invoiced.",
        "tags": [
          "invoices"
        ]
      },
      "get": {
        "operationId": "InvoicesController_byWorkOrder",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workOrderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            }
          },
          "404": {
            "description": "La OT no tiene factura (o no es de este taller)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Factura de una OT.",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/shops/{shopId}/invoices": {
      "get": {
        "operationId": "InvoicesController_list",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceViewPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Facturas del taller, de la más nueva a la más vieja.",
        "tags": [
          "invoices"
        ]
      }
    },
    "/v1/shops/{shopId}/invoices/{invoiceId}": {
      "get": {
        "operationId": "InvoicesController_get",
        "parameters": [
          {
            "name": "shopId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceView"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalle de una factura.",
        "tags": [
          "invoices"
        ]
      }
    },
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Cuenta provisionada con sus roles.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Perfil + roles del token actual (provisiona en la 1ª llamada).",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/roles": {
      "post": {
        "operationId": "MeController_addRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta con el rol añadido (o ya existente).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Añade un rol a la cuenta actual. Idempotente; rechaza courier/admin con 400.",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "MechMate API",
    "description": "Backend del ecosistema MechMate (ERP de talleres + marketplace de repuestos, RD). Dinero en centavos como string (`Cents`); fechas ISO-8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateShopDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Taller Hermanos Pérez",
            "minLength": 2,
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "enum": [
              "mechanic_shop",
              "parts_seller"
            ],
            "default": "mechanic_shop"
          }
        },
        "required": [
          "name"
        ]
      },
      "InviteMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "mecanico@gmail.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "mechanic",
              "advisor"
            ],
            "example": "mechanic"
          }
        },
        "required": [
          "email",
          "role"
        ]
      },
...
      "AddRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "mechanic",
              "seller",
              "customer"
            ],
            "description": "Rol a añadir. Solo mechanic|seller|customer (courier/admin se rechazan con 400)."
          }
        },
        "required": [
          "role"
        ]
      },
...
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "auth0_sub": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "kyc_status": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "auth0_sub",
          "email",
          "phone",
          "full_name",
          "status",
          "kyc_status",
          "roles",
          "created_at"
        ],
        "additionalProperties": false
      },
      "Role": {
        "type": "string",
        "enum": [
          "mechanic",
          "seller",
          "customer",
          "courier",
          "admin"
        ]
      },
      "ShopView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ShopType"
          },
          "commission_bps": {
            "type": "number"
          },
          "my_role": {
            "$ref": "#/components/schemas/ShopMemberRole"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "commission_bps",
          "my_role",
          "created_at"
        ],
        "additionalProperties": false
      },
      "ShopType": {
        "type": "string",
        "enum": [
          "mechanic_shop",
          "parts_seller"
        ]
      },
      "ShopMemberRole": {
        "type": "string",
        "enum": [
          "owner",
          "mechanic",
          "advisor"
        ]
      },
      "ShopMemberView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ShopMemberRole"
          },
          "status": {
            "type": "string",
            "enum": [
              "invited",
              "active"
            ]
          },
          "account_id": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "account_id",
          "email",
          "full_name",
          "created_at"
        ],
        "additionalProperties": false
      },
...
      "Cents": {
        "type": "string"
      },
...
      "WorkOrderItemType": {
        "type": "string",
        "enum": [
          "labor",
          "part"
        ]
      },
...
      "InvoiceView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "code": {
            "type": "string"
          },
          "ncf": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "issued",
              "paid",
              "voided"
            ]
          },
          "work_order_id": {
            "type": "string"
          },
          "work_order_code": {
            "type": "string"
          },
          "shop_name": {
            "type": "string"
          },
          "customer_name": {
            "type": "string"
          },
          "customer_document_id": {
            "type": "string",
            "nullable": true
          },
          "vehicle_description": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "subtotal_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "tax_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "total_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "issued_at": {
            "type": "string"
          },
          "paid_at": {
            "type": "string",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineView"
            }
          }
        },
        "required": [
          "id",
          "number",
          "code",
          "ncf",
          "status",
          "work_order_id",
          "work_order_code",
          "shop_name",
          "customer_name",
          "customer_document_id",
          "vehicle_description",
          "currency",
          "subtotal_cents",
          "tax_cents",
          "total_cents",
          "issued_at",
          "paid_at",
          "lines"
        ],
        "additionalProperties": false
      },
      "InvoiceLineView": {
        "type": "object",
        "properties": {
          "position": {
            "type": "number"
          },
          "type": {
            "$ref": "#/components/schemas/WorkOrderItemType"
          },
          "description": {
            "type": "string"
          },
          "part_number": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "string"
          },
          "unit_price_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "tax_rate_bps": {
            "type": "number"
          },
          "subtotal_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "tax_cents": {
            "$ref": "#/components/schemas/Cents"
          },
          "total_cents": {
            "$ref": "#/components/schemas/Cents"
          }
        },
        "required": [
          "position",
          "type",
          "description",
          "part_number",
          "quantity",
          "unit_price_cents",
          "tax_rate_bps",
          "subtotal_cents",
          "tax_cents",
          "total_cents"
        ],
        "additionalProperties": false
      },
...
      "InvoiceViewPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceView"
            }
          },
          "next_cursor": {
            "type": "string",
            "nullable": true,
            "description": "Pasar como `cursor` para la siguiente página; null si no hay más."
          }
        },
        "required": [
          "items",
          "next_cursor"
        ]
      }
    }
  }
}